requirements.txt
python-telegram-bot==21.3
yt-dlp>=2024.5.27
requests
PyYAML

# Optional (for search functionality)
google-api-python-client
isodate